backend/generate-from-template/requirements.txt
psycopg2-binary==2.9.9
anthropic==0.39.0
